WNB-181: Move remaining .js files in /store to .ts
