+++ src/store/auth/updateUserInfo/constants.ts
@@ -1,9 +1,3 @@
 export const UPDATE_USER_INFO_REQUEST = 'UPDATE_USER_INFO_REQUEST';
 export const UPDATE_USER_INFO_SUCCESS = 'UPDATE_USER_INFO_SUCCESS';
 export const UPDATE_USER_INFO_FAILURE = 'UPDATE_USER_INFO_FAILURE';
-
-export const INITIAL_STATE = {
-  loading: false,
-  loaded: false,
-  error: '',
-};
