+++ src/store/auth/updateUserInfo/reducer.ts
@@ -1,6 +1,6 @@
 import { handleActions } from 'redux-actions';
+
 import {
-  INITIAL_STATE,
   UPDATE_USER_INFO_REQUEST,
   UPDATE_USER_INFO_SUCCESS,
   UPDATE_USER_INFO_FAILURE,
@@ -10,7 +10,15 @@ import {
   handleLoad as handleRequest,
   handleLoadFailure as handleRequestFailure,
 } from '../../common/reducerHandlers';
+import { EntityInitialState } from '../../../types';
 
+const INITIAL_STATE = {
+  loading: false,
+  loaded: false,
+  // error: '',
+};
+
+// TODO: refactor to Typesafe actions in WNB-202
 export default handleActions(
   {
     [UPDATE_USER_INFO_REQUEST]: state => handleRequest(state),
@@ -21,7 +29,7 @@ export default handleActions(
   INITIAL_STATE
 );
 
-const handleRequestSuccess = state => {
+const handleRequestSuccess = (state: EntityInitialState) => {
   return {
     ...state,
     loading: false,
