+++ src/store/auth/user/reducer.ts
@@ -64,66 +64,3 @@ export const loadUserReducer = createReducer<EntityInitialState, UserActions>(
     }),
   }
 );
-
-// const handleRequestStart = (state, { payload }) => {
-//   return {
-//     ...state,
-//     loading: true,
-//     loaded: false,
-//     error: '',
-//   };
-// };
-//
-// // TODO: can we reduce duplication in this file?
-// const handleLoadUserSuccess = (
-//   state,
-//   { payload: { user, accessToken, message } }
-// ) => {
-//   return {
-//     ...state,
-//     profile: user,
-//     loaded: true,
-//     loading: false,
-//     accessToken,
-//     error: message, // message in payload indicates something's gone wrong (for now)
-//   };
-// };
-//
-// const handleLoadUserFailure = (state, { payload }) => {
-//   return {
-//     ...state,
-//     loaded: false,
-//     loading: false,
-//     error: R.isEmpty(payload) ? '' : payload,
-//     profile: null,
-//   };
-// };
-//
-// const handleLoadUserUnauthorized = state => {
-//   return {
-//     ...state,
-//     profile: null,
-//     loaded: true,
-//     loading: false,
-//     accessToken: null,
-//     // if user is unauthorized, we don't need any error message in store,
-//     // otherwise our forms will have this error message underneath whenever
-//     // an unauthorized user enters pages with these forms
-//     error: null,
-//   };
-// };
-//
-// const handleUpdateUserInfoSuccess = (state, { payload }) => {
-//   return {
-//     ...state,
-//     user: { ...payload },
-//     loading: false,
-//     loaded: true,
-//     error: '',
-//   };
-// };
-//
-// const handleSignout = state => ({
-//   ...state,
-//   profile: null,
-// });
