+++ src/store/organizations/reducer.ts
@@ -5,19 +5,21 @@ import {
   LOAD_ORGANIZATIONS_SUCCESS,
   LOAD_ORGANIZATIONS_FAILURE,
 } from './constants';
-import { handleLoad, handleLoadFailure } from '../common/reducerHandlers';
+import { handleLoad, handleLoadFailure } from '../common';
+import { EntityInitialState, EntityLoadSuccessPayload } from '../../types';
 
 const INITIAL_STATE = {
   loading: false,
   loaded: false,
-  error: '',
-  records: [],
+  // error: '',
+  // records: null,
 };
 
 export const organizationsReducer = handleActions(
   {
     [LOAD_ORGANIZATIONS_REQUEST]: state => handleLoad(state),
-    [LOAD_ORGANIZATIONS_SUCCESS]: (state, action) =>
+    // TODO: move organizations to Typesafe actions and get rid of 'any'
+    [LOAD_ORGANIZATIONS_SUCCESS]: (state, action: any) =>
       handleLoadSuccess(state, action),
     [LOAD_ORGANIZATIONS_FAILURE]: (state, action) =>
       handleLoadFailure(state, action),
@@ -25,12 +27,14 @@ export const organizationsReducer = handleActions(
   INITIAL_STATE
 );
 
-const handleLoadSuccess = (state, { payload }) => {
+const handleLoadSuccess = (
+  state: EntityInitialState,
+  { payload }: EntityLoadSuccessPayload
+) => {
   return {
     ...state,
     loading: false,
     loaded: true,
-    records: [...payload],
-    error: '',
+    records: payload && [...payload],
   };
 };
